+++ package.json
@@ -2,7 +2,7 @@
   "name": "cyberopsbot",
   "version": "1.0.0",
   "description": "",
-  "main": "CyberOpsBot.js",
+  "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
